# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ReturnType_typeofGetBooksOutputDto.parse_": {
				"items": {
					"properties": {
						"totalPages": {
							"type": "number",
							"format": "double"
						},
						"author": {
							"type": "string"
						},
						"summary": {
							"type": "string"
						},
						"title": {
							"type": "string"
						},
						"id": {
							"type": "string"
						}
					},
					"type": "object"
				},
				"type": "array",
				"description": "Obtain the return type of a function type"
			},
			"GetBooksOutputDto": {
				"$ref": "#/components/schemas/ReturnType_typeofGetBooksOutputDto.parse_"
			},
			"ReturnType_typeofGetBookOutputDto.parse_": {
				"properties": {
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Obtain the return type of a function type"
			},
			"GetBookOutputDto": {
				"$ref": "#/components/schemas/ReturnType_typeofGetBookOutputDto.parse_"
			},
			"ReturnType_typeofPostBookInputDto.parse_": {
				"properties": {
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Obtain the return type of a function type"
			},
			"PostBookInputDto": {
				"$ref": "#/components/schemas/ReturnType_typeofPostBookInputDto.parse_"
			},
			"ReturnType_typeofPostUserOutputDto.parse_": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Obtain the return type of a function type"
			},
			"PostUserOutputDto": {
				"$ref": "#/components/schemas/ReturnType_typeofPostUserOutputDto.parse_"
			},
			"ReturnType_typeofPostUserInputDto.parse_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"login": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Obtain the return type of a function type"
			},
			"PostUserInputDto": {
				"$ref": "#/components/schemas/ReturnType_typeofPostUserInputDto.parse_"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "clean-archi-nodejs",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/books": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetBooksOutputDto"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostBookInputDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostBookInputDto"
							}
						}
					}
				}
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetBookOutputDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					},
					"400": {
						"description": "Invalid book id format"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUserOutputDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUserInputDto"
							}
						}
					}
				}
			}
		},
		"/users/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUserOutputDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"409": {
						"description": "Already exists"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUserInputDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Localhost"
		}
	]
}